完善登录和注册模块，前后端整合联调

+++ src/main/java/com/chris/controller/ApplyController.java
@@ -1,6 +1,8 @@
 package com.chris.controller;
 
+import java.text.DateFormat;
 import java.text.ParseException;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
@@ -20,6 +22,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -28,18 +31,22 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.alibaba.fastjson.JSONObject;
 import com.chris.model.Apply;
+import com.chris.model.HostHolder;
 import com.chris.model.User;
 import com.chris.service.ApplyService;
 import com.chris.service.UserService;
+import com.chris.util.ApplySystemUtils;
 
-//@CrossOrigin(origins = "http://localhost:8080")
-@CrossOrigin
+@CrossOrigin(origins = "http://localhost:8080")
+//@CrossOrigin
 @Controller
 public class ApplyController {
 	@Autowired
 	private ApplyService applyService;
 	@Autowired
 	private UserService userService;
+	@Autowired
+	private HostHolder hostHolder;
 	
 	@RequestMapping(path = "/apply/add")
 	@ResponseBody
@@ -55,9 +62,13 @@ public class ApplyController {
              @RequestParam(value = "result", required = false) String result,
              HttpServletRequest request,
              HttpServletResponse response) throws ParseException {
+		User user = hostHolder.getUser();
+		//判断用户是否登录
+		if(user == null) {
+			return ApplySystemUtils.getJSONString(ApplySystemUtils.CODE_ERR, "请登录后操作");
+		}
 		Apply apply = new Apply();
 		apply.setUserId(Integer.valueOf(userId));
-		User user = userService.getUser(userId);
 		apply.setUsername(user.getUsername());
 		apply.setPeriodName(periodName);
 		apply.setCompanyName(companyName);
@@ -77,12 +88,16 @@ public class ApplyController {
 	@RequestMapping(path = "/apply/get")
 	@ResponseBody
 	public String getApply(Model model,
-			 @RequestParam("userId") int userId,
              @RequestParam(value = "periodName", required = false) String periodName,
              HttpServletRequest request,
              HttpServletResponse response) throws ParseException {
+		User user = hostHolder.getUser();
+		//判断用户是否登录
+		if(user == null) {
+			return ApplySystemUtils.getJSONString(ApplySystemUtils.CODE_ERR, "请登录后操作");
+		}
+		int userId = user.getUserId();
 		List<Apply> applies = applyService.selectByUserId(userId);
-		User user = userService.getUser(userId);
 		Map<String, Object> res = new HashMap<>();
 		res.put("userId", user.getUserId());
 		res.put("username", user.getUsername());
@@ -93,10 +108,10 @@ public class ApplyController {
 	}
 	
 	
+	
 	@RequestMapping(path = "/apply/update")
 	@ResponseBody
 	public String updateApply(Model model, 
-			 @RequestParam("userId") int userId,
 			 @RequestParam(value = "applyId") int applyId,
              @RequestParam(value = "periodName", required = false) String periodName,
              @RequestParam(value = "companyName", required = false) String companyName,
@@ -108,20 +123,24 @@ public class ApplyController {
              @RequestParam(value = "result", required = false) String result,
              HttpServletRequest request,
              HttpServletResponse response) throws ParseException {
+		User user = hostHolder.getUser();
+		//判断用户是否登录
+		if(user == null) {
+			return ApplySystemUtils.getJSONString(ApplySystemUtils.CODE_ERR, "请登录后操作");
+		}
 		Apply apply = new Apply();
 		apply.setApplyId(applyId);
 		if(periodName != null) apply.setPeriodName(periodName);
 		if(companyName != null) apply.setCompanyName(companyName);
-		if(applyDate != null) apply.setApplyDate(DateUtils.parseDate(applyDate, "yyyy-MM-dd"));
-		if(endDate != null) apply.setEndDate(DateUtils.parseDate(endDate, "yyyy-MM-dd"));
+		if(applyDate != null) apply.setApplyDate(DateUtils.parseDate(applyDate,"yyyy-MM-dd"));
+		if(endDate != null) apply.setEndDate(DateUtils.parseDate(endDate,"yyyy-MM-dd"));
 		if(recommend != null) apply.setRecommend(recommend);
 		if(status != null) apply.setStatus(status);
 		if(anticipate != null) apply.setAnticipate(anticipate);
 		if(result != null) apply.setResult(result);
 		applyService.updateApply(apply);
 		
-		List<Apply> applies = applyService.selectByUserId(userId);
-		User user = userService.getUser(userId);
+		List<Apply> applies = applyService.selectByUserId(user.getUserId());
 		Map<String, Object> res = new HashMap<>();
 		res.put("userId", user.getUserId());
 		res.put("mail", user.getMail());
@@ -135,8 +154,13 @@ public class ApplyController {
 			 @RequestParam(value = "applyId") int applyId,
              HttpServletRequest request,
              HttpServletResponse response) throws ParseException {
+		User user = hostHolder.getUser();
+		//判断用户是否登录
+		if(user == null) {
+			return ApplySystemUtils.getJSONString(ApplySystemUtils.CODE_ERR, "请登录后操作");
+		}
 		applyService.deleteApply(applyId);
-		return "OK";
+		return ApplySystemUtils.getJSONString(ApplySystemUtils.CODE_OK, "成功删除");
 	}
 	@RequestMapping(path = "/apply/removeList")
 	@ResponseBody
@@ -145,6 +169,11 @@ public class ApplyController {
 			 @RequestParam(value = "applyIds") String applyIds,
              HttpServletRequest request,
              HttpServletResponse response) throws ParseException {
+		User user = hostHolder.getUser();
+		//判断用户是否登录
+		if(user == null) {
+			return ApplySystemUtils.getJSONString(ApplySystemUtils.CODE_ERR, "请登录后操作");
+		}
 		List<Integer> idList = new ArrayList<>();
 		String[] idStrs = applyIds.split(",");
 		for(int i = 0; i < idStrs.length; i++) {
@@ -152,8 +181,6 @@ public class ApplyController {
 			idList.add(Integer.parseInt(tmp));
 		}
 		applyService.deleteApplyList(idList);
-		return "OK";
+		return ApplySystemUtils.getJSONString(ApplySystemUtils.CODE_OK, "成功批量删除");
 	}
-	
-	
 }
